Configure mockserver backend for health check additional headers test
