+++ testsuite/backend/httpbin.py
@@ -1,7 +1,5 @@
 """Httpbin implementation of Backend"""
 
-from functools import cached_property
-
 from testsuite.backend import Backend
 from testsuite.kubernetes import Selector
 from testsuite.kubernetes.client import KubernetesClient
@@ -13,24 +11,9 @@ class Httpbin(Backend):
     """Httpbin deployed in Kubernetes as Backend"""
 
     def __init__(self, cluster: KubernetesClient, name, label, image, replicas=1) -> None:
-        super().__init__()
-        self.cluster = cluster
-        self.name = name
-        self.label = label
+        super().__init__(cluster, name, label)
         self.replicas = replicas
         self.image = image
-
-        self.deployment = None
-        self.service = None
-
-    @property
-    def reference(self):
-        return {"group": "", "kind": "Service", "port": 8080, "name": self.name, "namespace": self.cluster.project}
-
-    @property
-    def url(self):
-        """URL for the httpbin service"""
-        return f"{self.name}.{self.cluster.project}.svc.cluster.local"
 
     def commit(self):
         match_labels = {"app": self.label, "deployment": self.name}
@@ -53,17 +36,3 @@ class Httpbin(Backend):
             ports=[ServicePort(name="http", port=8080, targetPort="api")],
         )
         self.service.commit()
-
-    def delete(self):
-        with self.cluster.context:
-            if self.service:
-                self.service.delete()
-                self.service = None
-            if self.deployment:
-                self.deployment.delete()
-                self.deployment = None
-
-    @cached_property
-    def port(self):
-        """Service port that httpbin listens on"""
-        return self.service.get_port("http").port
